Introduce nulim. Add protocol for (rolling) sum, average and median.
Previously, the code only tracked the number of integer bits and fractional
bits of an encoding.  With this commit, this gets refined in that each encoding
tracks the maximum permissible numerator (nulim), and the number of integer
bits is derived from that value.  The advantage of doing so is that upper
bounds can be derived much more precisely, potentially saving many bits when
computing a rolling sum or rolling average.

This commit also introduces a protocol for computing sums, averages, and
medians of values in a way that minimizes code duplication.  An implementation
only has to provide a function for computing the rolling sum and rolling
median, and all the other functions are derived from those.